requirements.txt
constructs>=10.0.0,<11.0.0
aws-cdk-lib==2.81.0
aws_cdk.aws_apigatewayv2_integrations_alpha